Refactor build tools
Shuffled around dependencies so that only packages necessary for
building and running the production web version are included in in the
regular "dependencies" list. Also moved around functions so that
build-prod.js doesn't end up importing a bunch of electron packages it
doesn't need by being linked to build-tools.
