+++ package.json
@@ -7,25 +7,25 @@
     "test": "test"
   },
   "dependencies": {
-    "@exponent/electron-cookies": "^2.0.0",
-    "electron-serve": "^0.4.1",
-    "express": "^4.17.1"
-  },
-  "devDependencies": {
     "@babel/cli": "^7.11.6",
     "@babel/core": "^7.8.4",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/preset-env": "^7.8.4",
-    "@types/chai": "^4.2.3",
-    "@types/mocha": "^5.2.7",
     "babel-runtime": "^6.26.0",
     "browserify": "^16.5.0",
     "core-js": "^3.6.4",
+    "express": "^4.17.1",
+    "fs-extra": "^9.0.1",
+    "npm-run": "^5.0.1"
+  },
+  "devDependencies": {
+    "@exponent/electron-cookies": "^2.0.0",
+    "@types/chai": "^4.2.3",
+    "@types/mocha": "^5.2.7",
     "electron": "^16.0.0",
     "electron-installer-redhat": "^3.2.0",
     "electron-packager": "^15.1.0",
-    "fs-extra": "^9.0.1",
-    "npm-run": "^5.0.1"
+    "electron-serve": "^0.4.1"
   },
   "scripts": {
     "test": "echo \"to do this, launch a local server and run tests.html in your browser\" && exit 0",
